+++ app/reader/components/ui/welcome-section.tsx
@@ -1,0 +1,16 @@
+/**
+ * Welcome Section Component
+ * Extracted from dashboard for modularity
+ */
+
+export function WelcomeSection() {
+  return (
+    <div className="mb-8">
+      <h1 className="text-3xl font-bold text-gray-900 mb-1">Welcome back,</h1>
+      <h2 className="text-3xl font-bold text-gray-900 mb-2">Sarah!</h2>
+      <p className="text-gray-600">
+        Discover trusted health information tailored for you.
+      </p>
+    </div>
+  );
+}
